OP-3021 - refactored base class into abstract class
Moved to api.plugin to make it clearer.
